Move ViewFormulaAccordion code to SelectFormula
